refactor: #82 - move constans, TextNode and serviceWorker to ts
